requirements.txt
torch>=2.0.0
numpy>=1.24.0
einops>=0.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
scipy>=1.10.0
networkx>=3.1
pandas>=2.0.0
plotly>=5.14.0

